os-user: cleanup
